src/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pytest==8.3.3
mysql-connector-python==8.0.23

